+++ src/main/java/com/lukasz/employee/EmployeeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-public interface EmployeeRepository extends CrudRepository<Employee, Integer> {
-    List<Employee> findByParking_ParkingId(Integer parkingId);
+public interface EmployeeRepository extends CrudRepository<Employee, Long> {
+    List<Employee> findByParking_ParkingId(Long parkingId);
 }
