+++ src/main/java/com/lukasz/parking/ParkingRepository.java
@@ -2,5 +2,5 @@ package com.lukasz.parking;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ParkingRepository extends CrudRepository<Parking, Integer> {
+public interface ParkingRepository extends CrudRepository<Parking, Long> {
 }
